requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
daphne>=4.0,<5.0
celery>=5.0,<6.0
grpcio>=1.60,<2.0
grpcio-tools>=1.60,<2.0
pydantic>=2.0,<3.0
pika>=1.3,<2.0
msal>=1.28,<2.0
pyjwt[crypto]>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
pgvector>=0.3,<1.0
